Sonoma (v14.x.x) compatibility
It seems like the logs in Sonoma (Mac OS v14.x.x) behave differently when the camera stream is on or off.  Updated the script so it looks for the 'AVFCapture' entry in the log and detects if the camera is on using 'AVCaptureSession_Tundra startRunning' and off using 'AVCaptureSession_Tundra stopRunning' entries which appear to be reliably consistent for both the built-in cam or a 3rd party cam.